Newer version of rubygems allows to bypass force
This has not been tested in respect to this project.
I used this project as a very helpful guidline in building RPMs and wanted to share what I learned.
I was unhappy with the need need to use the --force flag and found that in rubgems 2.4 --build-root was added
Hope this helps. Thanks for the information that this repo has provided me.