Add more files to ignore and doc
These are some files the Foreman project has found in various gems which
are common.

appveyor.yml is configuration for https://www.appveyor.com/
coveralls.yml is configuration for https://coveralls.io/

Both are only used in CI pipelines and don't make sense in RPMs.

.github is configuration for various Github related matters. Technically
this can contain CONTRIBUTING.md so it's not perfect. The alternative is
only exclude .github/workflows to exclude Github Actions, but we found
that it's generally not a problem to exclude .github.